✨ 新增超级用户与管理员帮助模板 (#1655)
