+++ resources/template/help/main.html
@@ -1,0 +1,62 @@
+<!DOCTYPE html>
+<html lang="zh-cn">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>test</title>
+    <!-- <link rel="stylesheet" href="./res/font-awesome/css/font-awesome.min.css"> -->
+    <link rel="stylesheet" href="main.css">
+</head>
+
+<body>
+    <div class="wrapper">
+        <div class="main">
+            <div class="title">
+                {{data.nickname}}的{{data.help_name}}帮助
+            </div>
+            <div class="items-border">
+            {% for plugin in data['plugin_list'] %}
+                <div class="items">
+                    <div class="item-title">
+                        {{plugin.name}}
+                    </div>
+                    <div class="item-des">
+                        简介: {{plugin.description}}
+                    </div>
+                    <div class="item-usage">
+                        <p class="usage-title">用法:</p>
+                        {{plugin.usage}}
+                    </div>
+                </div>
+            {% endfor %}
+            </div>
+            <!-- <div class="items">
+                <div class="item-title">
+                    刷屏禁言
+                </div>
+                <div class="item-des">
+                    简介: 刷屏禁言相关操作
+                </div>
+                <div class="item-usage">
+                    用法:<br>
+                    指令：<br>
+                    设置刷屏: 查看当前设置<br>
+                    -c [count]: 检测最大次数<br>
+                    -t [time]: 规定时间内<br>
+                    -d [duration]: 禁言时长<br>
+                    示例:<br>
+                        设置刷屏 -c 10: 设置最大次数为10<br>
+                        设置刷屏 -t 100 -d 20: 设置规定时间和禁言时长<br>
+                        设置刷屏 -d 10: 设置禁言时长为10<br>
+                    * 即 X 秒内发送同样消息 N 次，禁言 M 分钟 *
+                </div>
+            </div> -->
+        </div>
+    </div>
+</body>
+<script type="text/javascript" src="main.js">
+</script>
+
+</html>
